+++ ailanguage/batch_language_translation_result.go
@@ -5,8 +5,8 @@
 // Language API
 //
 // OCI Language Service solutions can help enterprise customers integrate AI into their products immediately using our proven,
-//     pre-trained and custom models or containers, without a need to set up an house team of AI and ML experts.
-//     This allows enterprises to focus on business drivers and development work rather than AI and ML operations, which shortens the time to market.
+// pre-trained and custom models or containers, without a need to set up an house team of AI and ML experts.
+// This allows enterprises to focus on business drivers and development work rather than AI and ML operations, which shortens the time to market.
 //
 
 package ailanguage
@@ -17,27 +17,24 @@ import (
 	"strings"
 )
 
-// EntityDocument The document details for entities detect call.
-type EntityDocument struct {
+// BatchLanguageTranslationResult Result of translation call.
+type BatchLanguageTranslationResult struct {
 
-	// Document unique identifier defined by the user.
-	Key *string `mandatory:"true" json:"key"`
+	// List of succeeded document response.
+	Documents []TranslationDocumentResult `mandatory:"true" json:"documents"`
 
-	// Document text for detect entities.
-	Text *string `mandatory:"true" json:"text"`
-
-	// Language code as per ISO 639-1 (https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes) standard.
-	LanguageCode *string `mandatory:"false" json:"languageCode"`
+	// List of failed document response.
+	Errors []DocumentError `mandatory:"false" json:"errors"`
 }
 
-func (m EntityDocument) String() string {
+func (m BatchLanguageTranslationResult) String() string {
 	return common.PointerString(m)
 }
 
 // ValidateEnumValue returns an error when providing an unsupported enum value
 // This function is being called during constructing API request process
 // Not recommended for calling this function directly
-func (m EntityDocument) ValidateEnumValue() (bool, error) {
+func (m BatchLanguageTranslationResult) ValidateEnumValue() (bool, error) {
 	errMessage := []string{}
 
 	if len(errMessage) > 0 {