+++ ailanguage/change_model_compartment_details.go
@@ -5,8 +5,8 @@
 // Language API
 //
 // OCI Language Service solutions can help enterprise customers integrate AI into their products immediately using our proven,
-//     pre-trained and custom models or containers, without a need to set up an house team of AI and ML experts.
-//     This allows enterprises to focus on business drivers and development work rather than AI and ML operations, which shortens the time to market.
+// pre-trained and custom models or containers, without a need to set up an house team of AI and ML experts.
+// This allows enterprises to focus on business drivers and development work rather than AI and ML operations, which shortens the time to market.
 //
 
 package ailanguage
@@ -17,27 +17,21 @@ import (
 	"strings"
 )
 
-// TextClassificationDocument The document details for test classification detect call.
-type TextClassificationDocument struct {
+// ChangeModelCompartmentDetails Details for changing the compartment of an Model.
+type ChangeModelCompartmentDetails struct {
 
-	// Document Unique Identifier.
-	Key *string `mandatory:"true" json:"key"`
-
-	// Document text for detect text classification.
-	Text *string `mandatory:"true" json:"text"`
-
-	// Language code as per ISO 639-1 (https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes) standard.
-	LanguageCode *string `mandatory:"false" json:"languageCode"`
+	// The OCID (https://docs.cloud.oracle.com/Content/General/Concepts/identifiers.htm) of the compartment into which the resource should be moved.
+	CompartmentId *string `mandatory:"true" json:"compartmentId"`
 }
 
-func (m TextClassificationDocument) String() string {
+func (m ChangeModelCompartmentDetails) String() string {
 	return common.PointerString(m)
 }
 
 // ValidateEnumValue returns an error when providing an unsupported enum value
 // This function is being called during constructing API request process
 // Not recommended for calling this function directly
-func (m TextClassificationDocument) ValidateEnumValue() (bool, error) {
+func (m ChangeModelCompartmentDetails) ValidateEnumValue() (bool, error) {
 	errMessage := []string{}
 
 	if len(errMessage) > 0 {