+++ ailanguage/work_request_log.go
@@ -5,8 +5,8 @@
 // Language API
 //
 // OCI Language Service solutions can help enterprise customers integrate AI into their products immediately using our proven,
-//     pre-trained and custom models or containers, without a need to set up an house team of AI and ML experts.
-//     This allows enterprises to focus on business drivers and development work rather than AI and ML operations, which shortens the time to market.
+// pre-trained and custom models or containers, without a need to set up an house team of AI and ML experts.
+// This allows enterprises to focus on business drivers and development work rather than AI and ML operations, which shortens the time to market.
 //
 
 package ailanguage
@@ -17,27 +17,24 @@ import (
 	"strings"
 )
 
-// SentimentsDocument The document details for sentiments detect call.
-type SentimentsDocument struct {
+// WorkRequestLog A log message from the execution of a work request.
+type WorkRequestLog struct {
 
-	// Document Unique Identifier.
-	Key *string `mandatory:"true" json:"key"`
+	// Human-readable log message.
+	Message *string `mandatory:"true" json:"message"`
 
-	// Document text for detect sentiments.
-	Text *string `mandatory:"true" json:"text"`
-
-	// Language code as per ISO 639-1 (https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes) standard.
-	LanguageCode *string `mandatory:"false" json:"languageCode"`
+	// The time the log message was written. An RFC3339 formatted datetime string
+	Timestamp *common.SDKTime `mandatory:"true" json:"timestamp"`
 }
 
-func (m SentimentsDocument) String() string {
+func (m WorkRequestLog) String() string {
 	return common.PointerString(m)
 }
 
 // ValidateEnumValue returns an error when providing an unsupported enum value
 // This function is being called during constructing API request process
 // Not recommended for calling this function directly
-func (m SentimentsDocument) ValidateEnumValue() (bool, error) {
+func (m WorkRequestLog) ValidateEnumValue() (bool, error) {
 	errMessage := []string{}
 
 	if len(errMessage) > 0 {